rename ProtocolNavBar to EditNavigation
