+++ protocol-designer/src/molecules/LiquidButton/__tests__/LiquidButton.test.tsx
@@ -3,7 +3,7 @@ import { fireEvent, screen } from '@testing-library/react'
 
 import { i18n } from '../../../assets/localization'
 import { renderWithProviders } from '../../../__testing-utils__'
-import { LiquidButton } from '../LiquidButton'
+import { LiquidButton } from '../../../molecules/LiquidButton'
 
 import type { ComponentProps } from 'react'
