+++ protocol-designer/src/molecules/index.ts
@@ -2,5 +2,6 @@ export * from './CheckboxExpandStepFormField'
 export * from './CheckboxStepFormField'
 export * from './DropdownStepFormField'
 export * from './InputStepFormField'
+export * from './LiquidButton'
 export * from './ToggleExpandStepFormField'
 export * from './ToggleStepFormField'
