+++ protocol-designer/src/organisms/index.ts
@@ -8,6 +8,7 @@ export * from './ConfirmDeleteStagingAreaModal'
 export * from './DefineLiquidsModal'
 export * from './DisabledScreen'
 export * from './EditInstrumentsModal'
+export * from './EditNavigation'
 export * from './EditNickNameModal'
 export * from './EditProtocolMetadataModal'
 export * from './FileUploadMessagesModal/'
@@ -17,10 +18,9 @@ export * from './Kitchen'
 export * from './KnowledgeLink'
 export * from './LabwareOnDeck'
 export * from './LabwareUploadModal'
-export * from './NavigationBar'
+export * from './Navigation'
 export * from './PipetteInfoItem'
 export * from './Portal'
-export * from './ProtocolNavBar'
 export * from './SelectWellsModal'
 export * from './SettingsIcon'
 export * from './SlotDetailsContainer'
